enhancement #38: decompose math_constuction_building serivce
