+++ lib/widgets/math_constructions/index.dart
@@ -1,0 +1,11 @@
+export 'frac.dart';
+export 'exp.dart';
+export 'sqrt.dart';
+export 'integral.dart';
+export 'named_function.dart';
+export 'undifinit_integral.dart';
+export 'backets.dart';
+export 'limit.dart';
+export 'log.dart';
+export 'abs.dart';
+export 'derevative.dart';
